chore: Move BasicUser and SetupMFA to parsers and StatisticData to serializers
